Clean up controller logic
